Only process the output size < limit
Process a large output (700M+) may cause OOM in CI server and/or TRSS.
Update the logic to only process the output < limit. If the output
> limit, record the error and display in UI. The limit is set to 50M for
now as the regular build output is ~3M. We can adjust the limit as
needed.

This PR also includes the following changes:
- remove try catch in the lower level (LogStream) and the exception will
be caught at the higher level (BuildProcessor), so that we can store the
exception into DB
- add AlertMsg in UI
- add short sleep after processing each build to avoid overloading CI
server
- add log for each CI server request with [CIServerRequest] (for
debugging)
- reduce max memory from 6G to 4G in dev env (to match with
production env)

Signed-off-by: lanxia <lan_xia@ca.ibm.com>